Refactor member dropdown
